Pulls EXIF data from file to detect photospheres and avoid downloading twice
Uses exiv2 (now required to be installed on the server) to pull
information about the image, detect if it's a photosphere or not, and
only displays the link for photosphere images.

Passes in the relevant data to the photosphere object to speed things
up.

Also urlencodes the url path now because chrome was getting a syntax
error in photos that have urls with single quotes (apostrophes) in them.

Also uses the relative path instead of the file_path() to avoid the
cache buster. This way large images aren't downloaded again and again
with frequent viewings.